Separate icon check and radio components
