+++ lib/components/src/icon-radio/icon-radio.scss
@@ -1,0 +1,59 @@
+@use "@material/ripple";
+
+:host {
+  .mdc-radio__background {
+    position: absolute;
+    top: 8px;
+    right: 8px;
+  }
+}
+:host([width="fill"]) {
+  width: 100%;
+}
+.container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  text-align: center;
+  row-gap: 8px;
+  width: var(--width);
+  height: var(--height);
+  border: solid 2px var(--mdc-theme-border);
+  border-radius: 8px;
+  color: var(--mdc-theme-text-primary-on-background);
+  @include ripple.surface;
+  @include ripple.radius-unbounded;
+  @include ripple.states;
+}
+.container::after,
+.container::before {
+  border-radius: 8px;
+}
+.container:hover {
+  cursor: pointer;
+}
+
+.checked {
+  --mdc-theme-border: #59cecd;
+  border: solid 2px var(--mdc-theme-border);
+  background-color: var(--mdc-theme-surface-primary-highlight);
+}
+[name="icon"]::slotted(*) {
+  font-size: 32px;
+}
+[name="text"]::slotted(*) {
+  color: var(--mdc-theme-text-secondary-on-background);
+  font-family: var(--mdc-typography-subtitle1-font-family);
+  font-size: var(--mdc-typography-subtitle1-font-size);
+  font-weight: 400;
+  line-height: var(--mdc-typography-subtitle1-line-height);
+  letter-spacing: 0.4px;
+}
+
+.mdc-radio__native-control:enabled:checked
+  + .mdc-radio__background
+  .mdc-radio__outer-circle,
+.mdc-radio__inner-circle {
+  border-color: var(--mdc-theme-secondary, #018786);
+}
